+++ bootstrap/providers.php
@@ -11,4 +11,5 @@ return [
     // localizan por RUTA, no importando clases.
     App\Modules\Core\Providers\CoreServiceProvider::class,
     App\Modules\Identity\Providers\IdentityServiceProvider::class,
+    App\Modules\Creator\Providers\CreatorServiceProvider::class,
 ];